feat: add Clock widget support; enhance widget rendering and format handling in NewTab component
